backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
motor==3.3.2
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.25.0
